corporation id in character and market orders

+++ Updater.cpp
@@ -31,7 +31,9 @@
 namespace Evernus
 {
     void Updater::performVersionMigration(const CacheTimerRepository &cacheTimerRepo,
-                                          const Repository<Character> &characterRepo) const
+                                          const Repository<Character> &characterRepo,
+                                          const Repository<MarketOrder> &characterOrderRepo,
+                                          const Repository<MarketOrder> &corporationOrderRepo) const
     {
         QSettings settings;
 
@@ -45,13 +47,22 @@ namespace Evernus
 
         if (majorVersion == 0)
         {
-            if (minorVersion < 4)
+            if (minorVersion < 5)
             {
-                if (minorVersion < 3)
-                    settings.setValue(PriceSettings::autoAddCustomItemCostKey, false);
+                if (minorVersion < 4)
+                {
+                    if (minorVersion < 3)
+                        settings.setValue(PriceSettings::autoAddCustomItemCostKey, false);
 
-                cacheTimerRepo.exec(QString{"DROP TABLE %1"}.arg(cacheTimerRepo.getTableName()));
-                cacheTimerRepo.create(characterRepo);
+                    cacheTimerRepo.exec(QString{"DROP TABLE %1"}.arg(cacheTimerRepo.getTableName()));
+                    cacheTimerRepo.create(characterRepo);
+                }
+
+                characterRepo.exec(QString{"ALTER TABLE %1 ADD COLUMN corporation_id INTEGER NOT NULL DEFAULT 0"}.arg(characterRepo.getTableName()));
+                characterOrderRepo.exec(QString{"ALTER TABLE %1 ADD COLUMN corporation_id INTEGER NOT NULL DEFAULT 0"}.arg(characterOrderRepo.getTableName()));
+                corporationOrderRepo.exec(QString{"ALTER TABLE %1 ADD COLUMN corporation_id INTEGER NOT NULL DEFAULT 0"}.arg(corporationOrderRepo.getTableName()));
+
+                QMessageBox::information(nullptr, tr("Update"), tr("This update requires re-importing all data.\nPlease click on \"Import all\" after the update."));
             }
         }
     }
